backend/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
pypdf>=3.17.4
pdf2image>=1.16.3
Pillow>=10.2.0
python-multipart>=0.0.6
pydantic>=2.5.2
python-dotenv>=1.0.0
gunicorn>=21.2.0
werkzeug>=3.0.1
APScheduler>=3.10.4
